+++ cloudant-http/src/main/java/com/cloudant/http/interceptors/CookieInterceptor.java
@@ -25,10 +25,17 @@ import org.apache.commons.io.IOUtils;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UnsupportedEncodingException;
+import java.net.CookieManager;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.net.URLEncoder;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -51,8 +58,8 @@ public class CookieInterceptor implements HttpConnectionRequestInterceptor,
     private final static Logger logger = Logger.getLogger(CookieInterceptor.class
             .getCanonicalName());
     private final byte[] sessionRequestBody;
-    private String cookie = null;
-    private boolean shouldAttemptCookieRequest = true;
+    private final CookieManager cookieManager = new CookieManager();
+    private final AtomicBoolean shouldAttemptCookieRequest = new AtomicBoolean(true);
 
     /**
      * Constructs a cookie interceptor.
@@ -79,86 +86,125 @@ public class CookieInterceptor implements HttpConnectionRequestInterceptor,
 
         HttpURLConnection connection = context.connection.getConnection();
 
-        if (shouldAttemptCookieRequest) {
-            if (cookie == null) {
-                cookie = getCookie(connection.getURL(), context);
+        // First time we will have no cookies
+        if (cookieManager.getCookieStore().getCookies().isEmpty() && shouldAttemptCookieRequest
+                .get()) {
+            if (!requestCookie(connection.getURL(), context)) {
+                // Requesting a cookie failed, set a flag if we failed so we won't try again
+                shouldAttemptCookieRequest.set(false);
             }
-            connection.setRequestProperty("Cookie", cookie);
         }
 
+        if (shouldAttemptCookieRequest.get()) {
+
+            // Debug logging
+            if (logger.isLoggable(Level.FINEST)) {
+                logger.finest("Attempt to add cookie to request.");
+                logger.finest("Cookies are stored for URIs: " + cookieManager.getCookieStore()
+                        .getURIs());
+            }
+
+            // Apply any saved cookies to the request
+            try {
+                Map<String, List<String>> requestCookieHeaders = cookieManager.get(connection
+                        .getURL().toURI(), connection.getRequestProperties());
+                for (Map.Entry<String, List<String>> requestCookieHeader :
+                        requestCookieHeaders.entrySet()) {
+                    List<String> cookies = requestCookieHeader.getValue();
+                    if (cookies != null && !cookies.isEmpty()) {
+                        connection.setRequestProperty(requestCookieHeader.getKey(),
+                                listToSemicolonSeparatedString(cookies));
+                    } else {
+                        logger.finest("No cookie values to set.");
+                    }
+                }
+            } catch (IOException e) {
+                logger.log(Level.SEVERE, "Failed to read request properties", e);
+            } catch (URISyntaxException e) {
+                logger.log(Level.SEVERE, "Failed to convert request URL to URI for cookie " +
+                        "retrieval.");
+            }
+        }
         return context;
     }
 
     @Override
     public HttpConnectionInterceptorContext interceptResponse(HttpConnectionInterceptorContext
                                                                       context) {
-        HttpURLConnection connection = context.connection.getConnection();
 
-        String cookieHeader = connection.getHeaderField("Set-Cookie");
-        if (cookieHeader != null) {
-            cookie = this.extractCookieFromHeaderValue(cookieHeader);
-            return context;
-        }
+        // Check if this interceptor is valid before attempting any kind of renewal
+        if (shouldAttemptCookieRequest.get()) {
 
-        try {
-            boolean renewCookie = false;
-            int statusCode = connection.getResponseCode();
+            HttpURLConnection connection = context.connection.getConnection();
 
-            if (statusCode == HttpURLConnection.HTTP_FORBIDDEN || statusCode == HttpURLConnection
-                    .HTTP_UNAUTHORIZED) {
-                InputStream errorStream = connection.getErrorStream();
-                try {
-                    switch (statusCode) {
-                        case HttpURLConnection.HTTP_FORBIDDEN: //403
-                            //check if it was an expiry case
-                            String errorString = IOUtils.toString(errorStream, "UTF-8");
-                            // Check using a regex to avoid dependency on a JSON library.
-                            // Note (?siu) flags used for . to also match line breaks and for
-                            // unicode
-                            // case insensitivity.
-                            if (!errorString.matches("(?siu)" +
-                                    ".*\\\"error\\\"\\s*:\\s*\\\"credentials_expired\\\".*")) {
-                                // Wasn't a credentials expired, throw exception
-                                HttpConnectionInterceptorException toThrow = new
-                                        HttpConnectionInterceptorException(errorString);
-                                // Set the flag for deserialization
-                                toThrow.deserialize = true;
-                                throw toThrow;
-                            } else {
-                                // Was expired - set boolean to renew cookie
+            // If we got a 401 or 403 we might need to renew the cookie
+            try {
+                boolean renewCookie = false;
+                int statusCode = connection.getResponseCode();
+
+                if (statusCode == HttpURLConnection.HTTP_FORBIDDEN || statusCode ==
+                        HttpURLConnection.HTTP_UNAUTHORIZED) {
+                    InputStream errorStream = connection.getErrorStream();
+                    try {
+                        // Get the string value of the error stream
+                        String errorString = IOUtils.toString(errorStream, "UTF-8");
+                        logger.log(Level.FINE, String.format(Locale.ENGLISH, "Intercepted " +
+                                "response %d %s", statusCode, errorString));
+                        switch (statusCode) {
+                            case HttpURLConnection.HTTP_FORBIDDEN: //403
+                                // Check if it was an expiry case
+                                // Check using a regex to avoid dependency on a JSON library.
+                                // Note (?siu) flags used for . to also match line breaks and for
+                                // unicode
+                                // case insensitivity.
+                                if (!errorString.matches("(?siu)" +
+                                        ".*\\\"error\\\"\\s*:\\s*\\\"credentials_expired\\\".*")) {
+                                    // Wasn't a credentials expired, throw exception
+                                    HttpConnectionInterceptorException toThrow = new
+                                            HttpConnectionInterceptorException(errorString);
+                                    // Set the flag for deserialization
+                                    toThrow.deserialize = true;
+                                    throw toThrow;
+                                } else {
+                                    // Was expired - set boolean to renew cookie
+                                    renewCookie = true;
+                                }
+                                break;
+                            case HttpURLConnection.HTTP_UNAUTHORIZED: //401
+                                // We need to get a new cookie
                                 renewCookie = true;
-                            }
-                            break;
-                        case HttpURLConnection.HTTP_UNAUTHORIZED: //401
-                            IOUtils.toString(errorStream, "UTF-8");
-                            // We need to get a new cookie
-                            renewCookie = true;
-                            break;
-                        default:
-                            break;
+                                break;
+                            default:
+                                break;
+                        }
+                    } finally {
+                        IOUtils.closeQuietly(errorStream);
                     }
-                } finally {
-                    IOUtils.closeQuietly(errorStream);
-                }
-            }
-
-            if (renewCookie) {
-                cookie = getCookie(connection.getURL(), context);
-                // Don't resend request, failed to get cookie
-                if (cookie != null) {
-                    context.replayRequest = true;
+                    if (renewCookie) {
+                        logger.finest("Cookie was invalid attempt to get new cookie.");
+                        boolean success = requestCookie(connection.getURL(), context);
+                        if (success) {
+                            // New cookie obtained, replay the request
+                            context.replayRequest = true;
+                        } else {
+                            // Didn't successfully renew, maybe creds are invalid
+                            context.replayRequest = false; // Don't replay
+                            shouldAttemptCookieRequest.set(false); // Set the flag to stop trying
+                        }
+                    }
                 } else {
-                    context.replayRequest = false;
+                    // Store any cookies provided on the response
+                    storeCookiesFromResponse(connection);
                 }
+            } catch (IOException e) {
+                logger.log(Level.SEVERE, "Error reading response code or body from request", e);
             }
-        } catch (IOException e) {
-            logger.log(Level.SEVERE, "Failed to get response code from request", e);
         }
         return context;
 
     }
 
-    private String getCookie(URL url, HttpConnectionInterceptorContext context) {
+    private boolean requestCookie(URL url, HttpConnectionInterceptorContext context) {
         try {
             URL sessionURL = new URL(String.format("%s://%s:%d/_session",
                     url.getProtocol(),
@@ -177,29 +223,28 @@ public class CookieInterceptor implements HttpConnectionRequestInterceptor,
 
 
             HttpURLConnection connection = conn.execute().getConnection();
-            String cookieHeader = connection.getHeaderField("Set-Cookie");
             int responseCode = connection.getResponseCode();
 
             if (responseCode / 100 == 2) {
 
                 if (sessionHasStarted(connection.getInputStream())) {
-                    return this.extractCookieFromHeaderValue(cookieHeader);
+                    return storeCookiesFromResponse(connection);
                 } else {
-                    return null;
+                    return false;
                 }
-
             } else {
                 InputStream errorStream = connection.getErrorStream();
                 try {
-                    // Consume the error stream to avoid leaking connections
-                    String error = IOUtils.toString(errorStream, "UTF-8");
-                    // Log the error stream content
-                    logger.fine(error);
+                    if (errorStream != null) {
+                        // Consume the error stream to avoid leaking connections
+                        String error = IOUtils.toString(errorStream, "UTF-8");
+                        // Log the error stream content
+                        logger.fine(error);
+                    }
                 } finally {
                     IOUtils.closeQuietly(errorStream);
                 }
                 if (responseCode == 401) {
-                    shouldAttemptCookieRequest = false;
                     logger.severe("Credentials are incorrect, cookie authentication will not be" +
                             " attempted again by this interceptor object");
                 } else if (responseCode / 100 == 5) {
@@ -212,7 +257,6 @@ public class CookieInterceptor implements HttpConnectionRequestInterceptor,
                             "Failed to get cookie from server, response code %s, " +
                                     "cookie authentication will not be attempted again",
                             responseCode);
-                    shouldAttemptCookieRequest = false;
                 }
             }
         } catch (MalformedURLException e) {
@@ -220,9 +264,9 @@ public class CookieInterceptor implements HttpConnectionRequestInterceptor,
         } catch (UnsupportedEncodingException e) {
             logger.log(Level.SEVERE, "Failed to encode cookieRequest body", e);
         } catch (IOException e) {
-            logger.log(Level.SEVERE, "Failed to read cookie response header", e);
+            logger.log(Level.SEVERE, "Failed to read cookie response error stream", e);
         }
-        return null;
+        return false;
     }
 
     private boolean sessionHasStarted(InputStream responseStream) throws IOException {
@@ -245,7 +289,33 @@ public class CookieInterceptor implements HttpConnectionRequestInterceptor,
 
     }
 
-    private String extractCookieFromHeaderValue(String cookieHeaderValue) {
-        return cookieHeaderValue.substring(0, cookieHeaderValue.indexOf(";"));
+    private boolean storeCookiesFromResponse(HttpURLConnection connection) {
+
+        // Store any cookies from the response in the CookieManager
+        try {
+            logger.finest("Storing cookie.");
+            cookieManager.put(connection.getURL().toURI(), connection.getHeaderFields());
+            return true;
+        } catch (IOException e) {
+            logger.log(Level.SEVERE, "Failed to read cookie response header", e);
+            return false;
+        } catch (URISyntaxException e) {
+            logger.log(Level.SEVERE, "Failed to convert request URL to URI for cookie storage.");
+            return false;
+        }
     }
+
+    private String listToSemicolonSeparatedString(List<String> cookieStrings) {
+        // RFC 6265 says multiple cookie pairs should be "; " separated
+        StringBuilder builder = new StringBuilder();
+        int index = 0; // Count from 0 since we will increment before comparing to size
+        for (String cookieString : cookieStrings) {
+            builder.append(cookieString);
+            if (++index != cookieStrings.size()) {
+                builder.append("; ");
+            }
+        }
+        return builder.toString();
+    }
+
 }