+++ CHANGES.md
@@ -1,3 +1,7 @@
+# Unreleased
+- [NEW] Enabled `reduce` and other reduce related parameters to be set when using
+  `MultipleRequestBuilder`.
+
 # 2.5.1 (2016-07-19)
 - [IMPROVED] Made the 429 response code backoff optional and configurable. To enable the backoff add
   an instance of a `Replay429Interceptor` with the desired number of retries and initial backoff: