+++ CHANGES.md
@@ -4,6 +4,9 @@
 - [FIX] Consumed response streams in `client.shutdown()` and `CookieInterceptor` to prevent
   connection leaks.
 - [NEW] Added functionality to remove attachment from document by attachment name.
+- [FIX] Issue authenticating with a proxy server when connecting to a HTTPS database server.
+- [FIX] Throw an `IllegalArgumentException` if using an unsupported proxy type.
+- [IMPROVED] Documentation for connecting to and authenticating with proxy servers.
 
 # 2.5.1 (2016-07-19)
 - [IMPROVED] Made the 429 response code backoff optional and configurable. To enable the backoff add